Use floating point 0 - 1 when dealing with rgb colors. All code updated except when dealing directly with vertices (i.e. in the bias lighting code).
Part #3 of 3.

+++ doomsday/plugins/jdoom/include/r_defs.h
@@ -77,7 +77,7 @@ typedef struct xsector_s {
     } planes[2];    // {floor, ceiling}
 
     float           origlight;
-    byte            origrgb[3];
+    float           origrgb[3];
     xgsector_t     *xg;
 
 } xsector_t;
